+++ verifier/verifier-app/src/Pages/Verified.jsx
@@ -1,0 +1,12 @@
+import React from "react";
+
+function Verified() {
+
+    return (
+        <div className="Verified">
+            <p>You have been verified!</p>
+        </div>
+    );
+}
+
+export default Verified;
